// src/lib/api/models/UserStatsDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { TagInfoDto } from './TagInfoDto';
export type UserStatsDto = {
    attemptCounts: number;
    averageScoreInPercentage: number;
    tagInfos: Array<TagInfoDto>;
};